backend/requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
gunicorn
motor==3.7.1
pymongo==4.14.1
python-jose==3.5.0
passlib[bcrypt]==1.7.4
email-validator==2.3.0
python-dotenv==1.1.1
pydantic==2.11.7
starlette==0.47.3
httpx==0.28.1
sentry-sdk==2.35.1
Jinja2==3.1.6
PyYAML==6.0.2
rich==14.1.0
python-multipart==0.0.20